// src-ui/core/ipc/generated/WebsocketServerConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MessagePostFormat } from "./MessagePostFormat";
import type { WsRole } from "./WsRole";

export type WebsocketServerConfig = { host: string, port: number, reportSelfMessage: boolean, enableForcePushEvent: boolean, heartInterval: number, path: string, role: WsRole, enable: boolean, name: string, messagePostFormat: MessagePostFormat, token: string, debug: boolean, };